requirements.txt
openai
langchain
python-dotenv
pyodbc --no-binary=pyodbc==4.0.35
fastapi
uvicorn[standard]
pydantic
azure-identity